Added submodule `approximate_factors`.
This submodule is ment to be a replacement for
the submodul with the same name of the `padic_extension`
branch of mclf, with the goal of making this module
independent from mclf.

The main difference with its antetype is that
an object of the class `ApproximateFactor` remembers
the polynomial it is a factor of.