First working version of `approximate_factorization`
The module `approximate_factorization` implements
the factorization of an arbitrary polynomial over
a number field into *approximate prime factors*
with respect to a p-adic valuation.

An *approximate prime factor* represents a prime
factor of the polynomial over the completion
(a padic number field), and coan compute
approximations of such a factor to arbitrary
precision.

The module `cusp_resolution` is in a WIP mode.