First working version of `cusp_resolution`
The main function, `resolve_cusp`, receives a
plane curve of degree d>=3 over a number field K,
and a p-adic valuation v_K on K. The assumption
is that the equation of the curve is primitive
with respect to v_K and hence defines a plane
model integral model. We also assume that the
special fiber has a cusp in normal form at the
origin (1:0:0).

The function computes a valued extension L/K and
a base change matrix T over L, so that the transformed
plane integral model is a semistable cubic in
Weierstrass normal form (plus the line at infinity).
This cubic is then the one-tail resulting from a
a resolution of the cusp over L.