New features: value and precision
The *precision* of an approximate prime factor
has been redefined as the measure of the accuracy
with which the root of this factor can be computed
from the current approximation.

Also, the *value* of an approximate prime factor
is the valuation of the root of this factor.

Some further small changes, and a test function.