Fixed the precision issue (and several bugs)
Using the new notion of 'precision' the search
for the correct factor of `f` (the equation for
alpha) has been greatly simplified and is now
(hopefully) failsafe. We compute alpha just to the
riht precision to be able to tell for sure whether
beta and gamma have positive valuation.

We also fixed several bugs. `resolve_cusp` now
seems to be stable and compute a resolution in
reasonable time over QQ and various quadratic
number fields.