Cusp resolution (#14)
* Some small changes

Among a few small changes, added a comment that the
output of `resolve_cusp` is not yet what it should
be.

* First working version of `approximate_factorization`

The module `approximate_factorization` implements
the factorization of an arbitrary polynomial over
a number field into *approximate prime factors*
with respect to a p-adic valuation.

An *approximate prime factor* represents a prime
factor of the polynomial over the completion
(a padic number field), and coan compute
approximations of such a factor to arbitrary
precision.

The module `cusp_resolution` is in a WIP mode.

* `approximate_solution` now works in the test case

Next step is to integrate it with `cusp_resolution`

* First working version of `cusp_resolution`

The main function, `resolve_cusp`, receives a
plane curve of degree d>=3 over a number field K,
and a p-adic valuation v_K on K. The assumption
is that the equation of the curve is primitive
with respect to v_K and hence defines a plane
model integral model. We also assume that the
special fiber has a cusp in normal form at the
origin (1:0:0).

The function computes a valued extension L/K and
a base change matrix T over L, so that the transformed
plane integral model is a semistable cubic in
Weierstrass normal form (plus the line at infinity).
This cubic is then the one-tail resulting from a
a resolution of the cusp over L.

* test suites for cusp resolution

We implement a test suite which produces random
curves and runs `cusp_resolution`.

With these test, we have found problems with
setting the desired precision. So the code
for `cusp_resolution` is still WIP.

* New features: value and precision

The *precision* of an approximate prime factor
has been redefined as the measure of the accuracy
with which the root of this factor can be computed
from the current approximation.

Also, the *value* of an approximate prime factor
is the valuation of the root of this factor.

Some further small changes, and a test function.

* Fixed the precision issue (and several bugs)

Using the new notion of 'precision' the search
for the correct factor of `f` (the equation for
alpha) has been greatly simplified and is now
(hopefully) failsafe. We compute alpha just to the
riht precision to be able to tell for sure whether
beta and gamma have positive valuation.

We also fixed several bugs. `resolve_cusp` now
seems to be stable and compute a resolution in
reasonable time over QQ and various quadratic
number fields.

* Improved the documentation of `approximate_factors`
a bit.

* Improve the documentation of `cusp_resolution`.